try to util gpu to max

+++ models_lightning.py
@@ -155,7 +155,7 @@ class SIGGRAPHGenerator(BaseColor):
 
     def training_step(self, data, batch_idx):
         labels, input_batch, name = data
-        outputs_probs = self(input_batch.type(torch.uint8))
+        outputs_probs = self(input_batch)
         loss = F.kl_div(F.log_softmax(outputs_probs, dim=1), labels.permute([0, 3, 1, 2]), log_target=True,
                         reduction="mean")
         self.log('train_loss', loss)
